Release 3.0
