+++ package.json
@@ -17,6 +17,7 @@
     "vue": "^2.6.14",
     "vue-axios": "^3.3.7",
     "vue-codemirror": "^4.0.6",
+    "vue-meta": "^2.4.0",
     "vuex": "^3.6.2"
   },
   "devDependencies": {
