+++ public/index.html
@@ -1,23 +1,20 @@
 <!DOCTYPE html>
 <html lang="es">
   <head>
-    <meta name="description" content="compilador web para c++ gratis y facil de usar, modo proyecto, clases y estandares hasta c++17, colaboracion remota y entrada de usuario " />
+
+    <meta name="author" content="Kevin Bohorquez" />
+    <meta name="keyword" content="c++, online compiler, free, easy, http library, curl, user data, fast, simple, c++20, assembly inspector, good performance"/>
+    <meta name="description" content="compilador web para c++ gratis y facil de usar, modo proyecto, clases y estandares hasta c++20, colaboracion remota y entrada de usuario."/>
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
     <link rel="icon" href="<%= BASE_URL %>favicon.ico">
-    <title>Knock Cpp</title>
+    <title>Kcompiler | online c++ ☀️</title>
   </head>
-  <body class="negro-apagado">
-    <style> 
-    /* si, soy un demonio */
-    .negro-apagado {
-      background-color: #121212;
-    }     
-    
-    </style>
+  <body>
     <noscript>
-      <strong>We're sorry but <%= htmlWebpackPlugin.options.title %> doesn't work properly without JavaScript enabled. Please enable it to continue.</strong>
+      <strong> pls enabled Javascript.</strong>
+        <strong> knock online c++ compiler with asm inspector in c++20, with http library! </strong>
     </noscript>
     <div id="app"></div>
     <!-- built files will be auto injected -->
