--- src/components/barra.vue
@@ -1,39 +1,0 @@
-<template>
-  <b-navbar toggeable="lg" class="navbar-horizontal negro-apagado" type="dark">
-    <!-- primera parte de la barra -->
-    <b-navbar-nav>
-      <b-navbar-brand>
-        <b-icon icon="exclude" variant="white"></b-icon>
-        <span class="font-weight-bold ml-3">Knock-CPP</span>
-      </b-navbar-brand>
-    </b-navbar-nav>
-
-    <!-- segunda parte de la barra -->
-    <b-navbar-nav class="ml-auto">
-      <b-nav-item :href="github"
-        ><b-icon icon="github" variant="white"></b-icon
-      ></b-nav-item>
-      <b-nav-item :href="paypal"
-        ><b-icon icon="heart-fill" variant="white"></b-icon
-      ></b-nav-item>
-    </b-navbar-nav>
-  </b-navbar>
-</template>
-
-<script>
-export default {
-  name: "barra",
-  data() {
-    return {
-      github: this.$store.state.github,
-      paypal: this.$store.state.paypal,
-    };
-  },
-};
-</script>
-
-<style scoped>
-.negro-apagado {
-  background-color: #121212;
-}
-</style>
