+++ src/components/modales/options.vue
@@ -6,10 +6,8 @@
       v-model="standar"
       title="el estandar con el que se compilara c++"
     >
-      <b-form-select-option value="c++11">c++11</b-form-select-option>
-      <b-form-select-option value="c++17">c++17</b-form-select-option>
-      <b-form-select-option value="c++14">c++14</b-form-select-option>
-      <b-form-select-option value="c++2a">c++2a</b-form-select-option>
+      <b-form-select-option v-for="(estandar, index) in estandares" :key="index" :value="estandar.value">{{estandar.name}}</b-form-select-option>
+
     </b-form-select>
     <b-form-select
       variant="primary"
@@ -17,9 +15,7 @@
       v-model="optimizar"
       title="el nivel de optimizacion a la hora de compilar, esto se vera en el codigo ensamblador"
     >
-      <b-form-select-option value="1">O1</b-form-select-option>
-      <b-form-select-option value="2">O2</b-form-select-option>
-      <b-form-select-option value="3">O3</b-form-select-option>
+      <b-form-select-option v-for="(nivel, index) in niveles" :key="index" :value="nivel">O{{nivel}}</b-form-select-option>
     </b-form-select>
     <b-button
       :disabled="usecurl == 'on'"
@@ -44,7 +40,7 @@
     >
       <b-icon icon="share-fill"></b-icon>
     </b-button>
-    
+
     <!-- panel de compartir -->
     <b-modal title="Compartir" id="share-link" hide-footer hide-header>
         <p style="cursor:pointer" @click="$bvModal.hide('share-link')">X</p>
@@ -55,7 +51,7 @@
 </template>
 
 <script>
-import { mapGetters } from "vuex";
+import { mapState} from "vuex";
 import share from "./share.vue"
 
 export default {
@@ -67,41 +63,42 @@ export default {
 
   methods: {
     download() {
-      this.$store.dispatch("download");
+      this.$store.dispatch("compile/download", this);
     },
     share() {
-      this.$store.dispatch("share");
+      this.$store.dispatch("codespaces/share");
     },
 
     getAssembly() {
-      this.$store.dispatch("getAssembly");
+      this.$store.dispatch("compile/getAssembly", this);
     },
   },
 
   computed: {
-    ...mapGetters["visibles"],
+    ...mapState('compile', ["estandares", "niveles"]),
+
     standar: {
       get() {
-        return this.$store.state.standar;
+        return this.$store.state.compile.compile.standar;
       },
       set(value) {
-        this.$store.commit("superUpdate", { type: "standar", data: value });
+        this.$store.state.compile.compile.standar = value;
       },
     },
     optimizar: {
       get() {
-        return this.$store.state.optimizar;
+        return this.$store.state.compile.compile.optimizar;
       },
       set(value) {
-        this.$store.commit("superUpdate", { type: "optimizar", data: value });
+        this.$store.state.compile.compile.optimizar = value;
       },
     },
     usecurl: {
       get() {
-        return this.$store.state.usecurl;
+        return this.$store.state.compile.usecurl;
       },
       set(value) {
-        this.$store.commit("superUpdate", { type: "curl", data: value });
+        this.$store.state.compile.usecurl = value
       },
     },
   },
