+++ src/main.js
@@ -1,7 +1,13 @@
 import Vue from 'vue'
 import App from './App.vue'
+import Vuex from 'vuex'
+import VueMeta from 'vue-meta'
 
-import store from './store'
+
+import base from './stores/base'
+import compile from './stores/compile'
+import notes from './stores/notes'
+import codespaces from "@/stores/codespaces";
 
 import {BootstrapVue, IconsPlugin, BootstrapVueIcons} from 'bootstrap-vue'
 import VueCodemirror from 'vue-codemirror'
@@ -17,17 +23,28 @@ const socket = io('https://lobster-app-tadp5.ondigitalocean.app',{
   reconnectionDelay: 10000
 })
 
-
 Vue.config.productionTip = false
-
 Vue.prototype.$socketio = socket;
 
+Vue.use(VueMeta)
 Vue.use(BootstrapVue)
 Vue.use(IconsPlugin)
 Vue.use(BootstrapVueIcons)
 Vue.use(VueCodemirror)
+Vue.use(Vuex);
+
+const store = new Vuex.Store({
+  modules: {
+    base,
+    compile,
+    notes,
+    codespaces
+  },
+});
+
 
 new Vue({
+  el: '#app',
   store,
-  render: h => h(App),
-}).$mount('#app')
+  render: h => h(App)
+})
